% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exphist.R
\name{insert}
\alias{insert}
\alias{insert.default}
\alias{insert.exphist}
\title{Insert}
\usage{
insert(x, val)

\method{insert}{exphist}(x, val)

\method{insert}{exphist}(x, val)
}
\arguments{
\item{x}{an exponential histogram}

\item{val}{a value to be inserted into the histogram}
}
\value{
the histogram holding the newly inserted value
}
\description{
The generic function \code{insert} stores a value into one of
the the buckets of the exponential histogram.
}
\details{
For exponential histograms, \code{\link{insert.exphist}} will be used.
For any other objects, \code{\link{insert.default}} will be used
to spawn an error message. Use \code{methods(insert)}.
}
\examples{
## Create the exp. hist.
eh <- exphist()
## Compute the value
insert(eh, 1)
}
