% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exphist.R
\name{print.exphist}
\alias{print.exphist}
\title{Print method for class 'exphist'}
\usage{
\method{print}{exphist}(x, ...)
}
\arguments{
\item{x}{an exponential histogram.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print method for class 'exphist'
}
