% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exphist.R
\name{size}
\alias{size}
\alias{size.default}
\alias{size.exphist}
\title{Size}
\usage{
size(x)

\method{size}{default}(x)

\method{size}{exphist}(x)
}
\arguments{
\item{x}{an exponential histogram}
}
\value{
the size of the buckets of the exponential histogram
}
\description{
The generic function \code{size} returns the total size of
the buckets of the exponential histogram.
}
\details{
For exponential histograms, \code{\link{size.exphist}} will be used.
For any other objects, \code{\link{size.default}} will be used
to spawn an error message. Use \code{methods(size)}.
}
\examples{
## Create the exp. hist.
eh <- exphist()
## Compute the value
size(eh)
}
