% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exphist.R
\name{value}
\alias{value}
\alias{value.default}
\alias{value.exphist}
\title{Value}
\usage{
value(x)

\method{value}{default}(x)

\method{value}{exphist}(x)
}
\arguments{
\item{x}{an exponential histogram}
}
\value{
the accumulated value for the histogram
}
\description{
The generic function \code{value} computes the accumulated value
in the buckets of the exponential histogram.
}
\details{
For exponential histograms, \code{\link{value.exphist}} will be used.
For any other objects, \code{\link{value.default}} will be used
to spawn an error message. Use \code{methods(value)}.
}
\examples{
## Create the exp. hist.
eh <- exphist()
## Compute the value
value(eh)
}
