% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exphist.R
\name{exphist}
\alias{exphist}
\title{Create an instance of a exponential histogram}
\usage{
exphist(winsize = 100, epsilon = 0.05, duration = NULL)
}
\arguments{
\item{winsize}{size of the sliding window}

\item{epsilon}{maximum relative error}

\item{duration}{time duration for timebased buckets}
}
\value{
an object of class 'exphist'
}
\description{
Create an instance of a exponential histogram
}
